+++ src/mainBot/Watchtower.java
@@ -151,7 +151,7 @@ public class Watchtower extends RobotPlayer{
         //Comms stuff
         Comms.updateSector(rc);
 
-        boolean currentHpThresh = rc.getHealth()/rc.getType().getMaxHealth(1) > 0.7;
+        boolean currentHpThresh = (double)rc.getHealth()/rc.getType().getMaxHealth(1) > 0.7;
         if (!currentHpThresh && aboveHpThresh) {
             rc.writeSharedArray(52, rc.readSharedArray(52) - 1);
         } else if (currentHpThresh && !aboveHpThresh) {
